experimenter timesheet support [no ci]
