+++ src/main/java/org/jlab/btm/presentation/controller/ajax/EditCcShiftInfo.java
@@ -1,0 +1,108 @@
+package org.jlab.btm.presentation.controller.ajax;
+
+import org.jlab.btm.business.service.OpShiftService;
+import org.jlab.btm.presentation.util.BtmParamConverter;
+import org.jlab.smoothness.business.exception.UserFriendlyException;
+import org.jlab.smoothness.business.util.ExceptionUtil;
+
+import javax.ejb.EJB;
+import javax.ejb.EJBAccessException;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.validation.ConstraintViolationException;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.sql.SQLException;
+import java.util.Date;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+/**
+ * @author ryans
+ */
+@WebServlet(name = "EditCcShiftInfo", urlPatterns = {"/ajax/edit-cc-shift-info"})
+public class EditCcShiftInfo extends HttpServlet {
+
+    private static final Logger logger = Logger.getLogger(
+            EditCcShiftInfo.class.getName());
+
+    @EJB
+    OpShiftService shiftService;
+
+    /**
+     * Handles the HTTP <code>POST</code> method.
+     *
+     * @param request  servlet request
+     * @param response servlet response
+     * @throws ServletException if a servlet-specific error occurs
+     * @throws IOException      if an I/O error occurs
+     */
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response)
+            throws ServletException, IOException {
+        String errorReason = null;
+
+        try {
+            if (request.getRequestedSessionId() != null && !request.isRequestedSessionIdValid()) {
+                throw new UserFriendlyException(
+                        "Your session has expired.  Please reload the page and relogin.");
+            }
+
+            Date startDayAndHour = BtmParamConverter.convertDayAndHour(request, "startDayAndHour");
+            String crewChief = request.getParameter("crewChief");
+            String operators = request.getParameter("operators");
+            String program = request.getParameter("program");
+            String programDeputy = request.getParameter("programDeputy");
+            String comments = request.getParameter("comments");
+
+            shiftService.editShift(startDayAndHour, crewChief, operators, program, programDeputy,
+                    comments);
+        } catch (EJBAccessException e) {
+            logger.log(Level.WARNING,
+                    "Unable to edit shift info due to access exception", e);
+            errorReason = "Access Denied";
+        } catch (UserFriendlyException e) {
+            logger.log(Level.INFO,
+                    "Unable to edit shift info: {0}",
+                    e.getMessage());
+            errorReason = e.getMessage();
+        } catch (Exception e) {
+            logger.log(Level.SEVERE,
+                    "Unable to edit shift info", e);
+            Throwable rootCause = ExceptionUtil.getRootCause(e);
+            if (rootCause instanceof SQLException) {
+                errorReason = "Database exception";
+            } else if (rootCause instanceof ConstraintViolationException) {
+                errorReason = "Constraint violation";
+            } else {
+                errorReason = "Something unexpected happened";
+            }
+        }
+
+        response.setContentType("text/xml");
+
+        PrintWriter pw = response.getWriter();
+
+        String xml;
+
+        if (errorReason == null) {
+            xml = "<response><span class=\"status\">Success</span></response>";
+        } else {
+            xml = "<response><span class=\"status\">Error</span><span "
+                    + "class=\"reason\">" + errorReason + "</span></response>";
+        }
+
+        pw.write(xml);
+
+        pw.flush();
+
+        boolean error = pw.checkError();
+
+        if (error) {
+            logger.log(Level.SEVERE, "PrintWriter Error");
+        }
+    }
+}
